# src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: openapi@inss.ch
  description: |-
    Forked and modified by International Network Support & Service - Glas.
    This document contains some more endpoints containing APIs that don't have only common formats like application/json and other commonly used formats, and contains also float numbers and Umlaute.
    This is to test & check if a tool can handle also those formats.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://inss.ch
  title: Swagger Petstore - OpenAPI 3.0
  version: 1.0.0
...
servers:
- url: /api/v3
...
paths:
...
  /oops:
    get:
      description: Produces sample error response.
      operationId: failingRequest
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
          description: Never returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              explode: false
              schema:
                type: string
              style: simple
        "304":
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              explode: false
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
          description: Bad request.
      summary: Always fails
      tags:
      - troublemaker
      x-accepts: application/json
      x-tags:
      - tag: troublemaker
      - tag: failing
...
components:
...
  schemas:
...
    RestError:
      description: The schema for all error responses.
      properties:
        status:
          description: The HTTP status code.
          example: 400
          format: int32
          readOnly: true
          title: Status
          type: integer
        error:
          description: The short error message.
          example: Bad Request
          readOnly: true
          title: Error
          type: string
        path:
          description: The path of the URL for this request.
          example: /api/owners
          format: uri
          readOnly: true
          title: Path
          type: string
        timestamp:
          description: The time the error occured.
          format: date-time
          readOnly: true
          title: Timestamp
          type: string
        message:
          description: The long error message.
          example: Request failed schema validation
          readOnly: true
          title: Message
          type: string
        trace:
          description: The stacktrace for this error.
          example: 'com.atlassian.oai.validator.springmvc.InvalidRequestException:
            ...'
          readOnly: true
          title: Trace
          type: string
      required:
      - error
      - message
      - path
      - schemaValidationErrors
      - status
      - timestamp
      title: REST Error
      type: object
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey